Parse outputs with only IPs (`traceroute -n`)
When using `traceroute -n`, names will not be printed (no reverse DNS
queries will be done) and only IPs will be printed.

Without `-n`: `<name> (<ip>) <rtt> <anno>`. With `-n`: `<ip> <rtt>
<anno>`.

Note that the source code of `loads` has now a lot of duplicated code.
This is why we have optional '[ASx]' and '(ip)' match with complicates
things. So the code should be refactored.

Fixes: #5