Refactoring parse code (remove some code replication)
Using the probe line concept. Each rtt in a probe line represents probes
of the same host (asn/name/ip). This allow us to simplify the code using
`continue` statements.

The regexp `RE_PROBE_IP` was readded so match is based on a regexp
instead of a simple string `startswith` match.